pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["thisisel <1997.esm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
python-dotenv = "*"
tortoise-orm = "0.17.4"
asyncpg = "*"
aiosqlite = "*"
fastapi-users = "*"
loguru = "^0.5.3"
aerich = "^0.5.3"
fastapi-pagination = "^0.7.4"

[tool.poetry.dev-dependencies]
pyflakes = "^2.3.1"
requests = "^2.25.1"
pytest = "^5.2"
pytest-asyncio = "^0.15.1"
pprintpp = "^0.4.0"
httpx = "^0.18.2"
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
run = "run:run"